feat: move erc20 transfer to its own shared crate
