+++ Cargo.toml
@@ -32,6 +32,7 @@ members = [
     "precompiles/preimage",
     "precompiles/balances-erc20",
     "precompiles/assets-erc20",
+    "precompiles/erc20-utils",
     "precompiles/verify-ecdsa-secp256k1-signature",
     "precompiles/verify-ecdsa-secp256r1-signature",
     "precompiles/verify-ecdsa-stark-signature",
@@ -307,6 +308,7 @@ pallet-hotfix-sufficients = { git = "https://github.com/paritytech/frontier.git"
 precompile-utils = { git = "https://github.com/paritytech/frontier.git", branch = "stable2407", default-features = false }
 
 # Local precompiles
+evm-erc20-utils = { path = "precompiles/erc20-utils", default-features = false }
 pallet-evm-precompile-democracy = { path = "precompiles/pallet-democracy", default-features = false }
 pallet-evm-precompile-batch = { path = "precompiles/batch", default-features = false }
 pallet-evm-precompile-balances-erc20 = { path = "precompiles/balances-erc20", default-features = false }
