+++ precompiles/erc20-utils/Cargo.toml
@@ -1,0 +1,25 @@
+[package]
+name = "evm-erc20-utils"
+authors = { workspace = true }
+description = "A Package containing ERC20 utility functions for EVM precompiles."
+edition = { workspace = true }
+version = "0.1.0"
+
+[dependencies]
+ethabi = { workspace = true }
+hex = { workspace = true, default-features = false }
+precompile-utils = { workspace = true }
+log = { workspace = true, default-features = false }
+
+# Substrate
+sp-core = { workspace = true }
+sp-std = { workspace = true }
+
+# Frontier
+fp-evm = { workspace = true }
+
+[dev-dependencies]
+
+[features]
+default = ["std"]
+std = ["ethabi/std", "hex/std", "fp-evm/std", "precompile-utils/std", "sp-core/std", "sp-std/std"]
